+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Drop support for *OpenLayers* <6.6
 Switch to the built-in TypeScript bindings in *OpenLayers* >=6.6
 Add basic WMS/TileWMS support
+Support `zIndex` in `RStyle`
 
 ### [1.0.4] 2021-07-10