+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Drop support for *OpenLayers* <6.6
 Switch to the built-in TypeScript bindings in *OpenLayers* >=6.6
 Add basic WMS/TileWMS support
 Support `zIndex` in `RStyle`
+Fix #13: Dynamic rendering in `RStyleArray`
 
 ### [1.0.4] 2021-07-10